demo: Add 'flex' mode to meshlets()
In addition to the existing options (what's one more boolean?), we now
support calling the flex algorithm with roughly 1/3rd of the triangle
budget as minimum and a split factor of 2.0, as well as a negative cone
weight.

It's perhaps more accurate to call this configuration "RT-focused", as
it's just one instance of flexible clusterization; having a tighter
bound on cluster sizes and a split factor without axis aligned weighting
can be beneficial for hierarchical clustering for example.