gltfpack: Warn when position quantization error is significant for any mesh
When float quantization is not used, we use the same quantization
settings for all meshes. This simplifies the flow as skins and materials
may need to change based on the quantization grid, and in some cases
eliminates micro gaps between meshes, but may result in reduced
precision compared to using per-mesh grids.

For now we detect this case and suggest increasing the number of bits
used for positions or switching to floating-point quantization.