clusterizer: Compute a corner of the mesh to be used for global flow
We need a specific point to anchor the overall traversal to, to ensure
the traversal stays local and expands in all sides equally. Currently
we start at an origin in mesh local space and do not control the flow;
it doesn't matter as much *where* we start, but the origin may be
embedded inside the mesh which makes the order less predictable.