clusterizer: Introduce triangle seed list
To maintain meshlet global flow, we need to start new meshlets in the
order that prioritizes smaller distances to the corner. However, to
maintain local flow we still have to minimize the live_triangle based
scoring we use right now. To avoid high algorithmic complexity we need
to maintain a list of triangles of a limited size that we will use to
evaluate these; for now, just add the triangle that's closest to the
corner to the list.