clusterizer: Refill seed list when the meshlet is full
When the next triangle does not fit into the meshlet, we analyze the
neighbors of each vertex and append the best triangles to the seed list.

This process is not precise: we do not filter out duplicate neighbors
between different vertices, and we do not sort the replaced triangles
perfectly, just taking the first available slot instead.