clusterizer: Use seed triangles to start new meshlets in some cases
Whenever the best triangle doesn't fit into the current meshlet we now
prune the seed list and select the best seed triangle accounrding to the
live+distance metric; this replaces the logic for using the meshlet
neighbors. That helps ensure the global flow of meshlets remains
simultaneously optimal from the liveness perspective and clustered
spatially, which reduces the chance of split meshlets down the line.

The integration is currently partial, as the algorithm structure doesn't
lend itself for a single natural place to add this logic; notably, if a
meshlet is split, we actually don't use it for seeding and don't
correctly select the starting triangle for the next meshlet, which
will be addressed in the future.