clusterizer: Reduce duplicate seeds from the same meshlet
For now we aren't filtering the seeds from the meshlet precisely, and we
may get duplicates: we select one neighbor triangle per vertex which
means we will likely see the same triangle as a neighbor of another
vertex in the same meshlet.

Using non-strict comparison for replacement mitigates this issue
somewhat, as it allows replacing the triangle with itself instead of
forcing the triangle to occupy another slot. This improves the flow by
giving more options to choose from during selection later.