clusterizer: Expose sphere bounds computation
The existing functions for computing cluster bounds,
meshopt_computeMeshletBounds and meshopt_computeClusterBounds, rely on
an internal computeBoundingSphere algorithm. This algorithm can be
useful in a few other cases:

- The cluster functions limit the input size to avoid allocations, as
  they need to reshape the inputs into a form that can be consumed by
  the internal function; sometimes it's convenient to compute a bounding
  sphere of what is essentially an unbounded point cloud
- For hierarchical simplification, it's useful to be able to compute a
  bounding sphere for an input set of spheres instead of just points.

This change exposes a basic implementation of both of the above with one
function that takes a set of points and, optionally, a per-point radius.
The current implementation uses computeBoundingSphere to compute the
center of the resulting sphere ignoring the radii; this results in
suboptimal output and will be changed separately, just as the extra
allocations that are unnecessary and will be removed.