clusterizer: Use radii for extremum points in computeBoundingSphere
In cases where we need to find extremum points or expand the sphere to
fit all points, we now incorporate a per-point radius into the
computation. This math is *not* fully correct for, specifically, the
initial bounding sphere and the sphere expansion as it doesn't fully
account for the possibility that the points may have very different
radii.

To simplify the code, we require that the radii are always specified
internally; using stride=0 allows to bypass that requirement.