clusterizer: Remove redundant radius computation
Now that computeBoundingSphere is correct, we no longer need the
adjustment to the radius to ensure all spheres are covered.

Note that the resulting radius is subject to some floating-point errors,
so we can't necessarily replace the removed code with assertions, as the
containment of the spheres is true only with some floating point
tolerance.