demo: Replace boundsMerge function with a more precise variant
Instead of an ad-hoc sphere merge that sometimes produces suboptimal
results we use meshopt_copmuteSphereBounds to compute a much tighter
sphere. This probably is not crucial in practice since this sphere is
not used for culling, but there might be cases where it's more efficient
to just keep a single sphere, and it's a convenient use of the
algorithm.