clusterizer: Increase max vertex limit to 256
Before this change, the maximum meshlet configuration that we allowed was 255v/512t; "255" was not a typo and reflected the internal implementation limit: we track the index of each vertex in the meshlet with a sentinel value 0xff to indicate vertices that haven't been added yet.

This is a surprising paper cut that nobody expects, and while 256v isn't dramatically different from 255v, it's still not ideal.

Upon testing, while in theory using a 2-byte lookup entry to allow -1 for sentinels technically requires more memory, in practice it's not a concern for anything except buildMeshletsScan, and even there while this requires more memory, the performance seems equivalent. If this becomes a problem in the future we can switch to a bitset.