indexcodec: Restructure decoding fast path for better codegen
This change mostly intends to break a compiler "optimization" that ends
up slowing this code, which merges tails of all 4 blocks as all 4 of
them have two pushEdgeFifo calls that look the same. In addition to a
less efficient control flow, on MSVC in particular this seems to break
several codegen optimizations, which also results in locally suboptimal
code, such as using `imul` to compute the store address for edgefifo.

Since writeTriangle always has a branch, it's actually more efficient to
place it to the end of the block anyway as it results in a more
streamlined control flow, and this also happens to make the structure
different enough from the remaining two blocks that tail merging doesn't
happen.

This makes MSVC decoding ~8% faster.