indexgenerator: Initial sketch of fuzzy remap
Existing generateVertexRemap variants only support binary vertex
equivalence; in some cases, welding with tolerance is preferred. This
change adds an interface and implementation sketch of
meshopt_generateVertexRemapFuzzy for that purpose.

While it's possible to specify this function to work on floating point
inputs and per-component tolerance, this runs into issues with
non-floating-point vertex components like bone indices that need to be
compared exactly, and in some cases the components might need to be
compared with vertex format specific functions, e.g. dot(normal1,
normal2) >= threshold. So the interface instead exposes a generic
function that compares two vertices, which will be called for all
vertices with the same or similar position - typically, vertex valence
is small so this is hopefully a reasonable compromise wrt performance.