indexgenerator: Implement hashing/comparison for tolerance=0
Instead of using generic binary hashing/comparison, we now use a proper
hash function of the floating-point position vector (which relies on
exact equivalence, and models negative zero properly), and use the
provided comparison callback to deduplicate vertices further.

When meshopt_generateVertexRemapFuzzy is called with tolerance=0, this
results in the correct implementation of deduplication: all vertices
with identical positions will be inserted into the hash table and only
the ones with unique attributes as defined by the callback will survive.

When tolerance is not 0, the implementation is not correct yet, as
different positions will hash into different buckets, preventing correct
deduplication. This will be implemented separately.