indexgenerator: Add a template overload for fuzzy remap
C API requires passing an extra context parameter into the comparison
callback to make it possible for that callback to reference the relevant
vertex state to compare.

When using C++, we can wrap the API with a templated wrapper which, in
addition to the usual index conversion, allows passing any callable
object transparently, including a lambda with captures.