indexgenerator: Initial scaffolding for thresholded hashing
When position tolerance is not 0, hashing the actual position will
result in similar positions placed in separate buckets, which often will
result in equivalent pairs not being detected.

Instead what should happen is that the position is hashed with reduced
precision eight times: for each axis, it is possible to truncate the
value precision such that any value in [v, v+tol] interval truncates to
either the same value, or to the same value + 1. This results in 8
values (2^3) for three axes, and searching in all of them should find
the equivalent vertex if it's present.

This change sets up the structure of the code to be able to do that, but
doesn't implement the actual truncation yet.