indexgenerator: Implement tolerance truncation
Instead of hashing the original floating point value as is, when
tolerance is specified we truncate it to the precision of the tolerance
(or rather to the binary exponent at which tolerance < 1). This, along
with a correctly applied offset, allows us to produce two integers such
that any floating point number in (v..v+tol) should hash to one of these
two.