simplifier: Adjust position hasher to handle negative zero
Before this change, if two vertices had positions that only differ in
the sign of zero in one of the components, the vertices would be treated
as separate.

This could lead to some seam vertices being classified as two border
vertices instead, which then allowed the simplifier to pull them in two
different directions, creating a small gap.

To fix this, we need to make sure negative zero hashes to 0, and the
positions are compared using floating point equality instead of bitwise
equality. If any vertices have NaN components, this will lead to them
not being merged into one, but these are going to produce NaN quadrics
anyway which will effectively lock simplification in the neighborhood
regardless of topology.