indexgenerator: Remove support for non-zero positional tolerance
Non-zero positional tolerance has two issues: it's much more expensive
to compute, and it's often wrong to use.

On the performance front, three different implementation strategies were
evaluated:

- Sorting with bounded lookahead (similar to sweep & prune)
- Hashing with truncation/offsets, as implemented here
- Building a bounding volume hierarchy until nodes are smaller than
  tolerance

Sorting was quickly discarded as it results in many cases that are
fundamentally inefficient, not due to sorting itself but due to the
window of the comparison required; a N*N regular grid will have to run N
N^2 loops to identify zero overlapping vertices if the sorting happens
to align with one of the axes.

Hashing and bounding volume hierarchy are harder to compare. Bounding
volume hierarchy can be significantly faster in some cases (e.g. 2-3x),
as it is more memory local, and can quickly narrow a search region to
the point where it will only contain a couple vertices. However, it is
significantly more fragile as it uses recursive data flow based on the
coordinate values, and it is easy to accidentally get into problematic
situations, including degenerate volumes and cases where performance is
actually multiple times slower. Additionally, this results in a
completely separate implementation vs the case where positions should be
compared with zero tolerance, which is bad for maintainability.

The hash implementation is generally "reasonable" wrt performance, but
on large meshes it can be >8x slower vs the zero tolerance version: we
need to perform 8 hash lookups, and the hash table is 8x larger so the
accesses are less likely to fit into the L3 cache.

A few potential fixes have been experimented with, including a separate
pre-filtering step to try to identify duplicate or redundant hash
samples, doing all 8 probes in parallel (using GPU-style batching where
the entire batch needs to be processed until no active elements remain),
and prefetching all 8 probe positions before the actual hash samples.

Separate pre-filtering step mostly doesn't help; parallel probing is up
to 2x faster in some cases but up to 50% slower for very large tables;
and prefetching does help by ~50% for larger tables, which doesn't fully
alleviate the issues.

Perhaps a more significant problem, but something that was discovered
after doing performance work, is that the mechanism of remap based
welding with positional tolerance is fundamentally prone to introducing
gaps: if a vertex has multiple clones with exactly the same position but
different normals, and has another vertex next to it with a normal that
is close to one of the normals, the source vertex may be "split" as we
will remap one of the wedges to move slightly, and the other one may
have no valid remap targets - thus a mesh that was previously watertight
no longer is.

This is not a problem if the welding is performed with zero attributes,
but between two use cases (zero positional tolerance + attribute
tolerance vs positional tolerance + zero attributes), the former is
believed to be prevalent. So based on all of this, it looks like the
best solution is to remove positional tolerance support for now.