indexgenrator: Rename generateVertexRemapFuzzy to Custom
Now that we no longer support position tolerance, it makes less sense to
call the function Fuzzy, as it by itself only performs exact
comparisons. While the main usecase would be to compare attributes with
tolerance, it's also possible to use it to generate vertex remap with
exact equality, for example if the behavior of existing functions wrt
padding is inconvenient.