indexanalyzer: Merge vcacheanalyzer & vfetchanalyzer together
For posterity, indexanalyzer.cpp consists of the contents of both
vcacheanalyzer.cpp and vfetchanalyzer.cpp. To preserve git blame
information accurately, this had to be done with two separate rename
commits that got merged together, and the CMake file was updated
separately to avoid complicating the merges.

A single, slightly larger, and thematically coherent file seems better
in this case than two separate files given that both algorithms are
rarely used and are mostly provided for convenience.

Any build scripts that compiled one or both of these two files manually
can replicate a change similar to this.