rasterizer: Add experimental meshopt_analyzeCoverage
Computing pixel coverage of a mesh can be useful in some cases: it
allows to roughly understand how the silhouette changes during
simplification, and the resulting coverage values can be useful for
distance culling based on projected bounding box, where instead of just
using the projection size, a view dependent coverage can be computed
instead.

The implementation uses the same rasterizer that is used for overdraw
analysis.