demo: Initial SAH cost calculation for clusters
To fully evaluate the cost of clusterization in context of clustered RT
we need to build a SAH optimized BVH over the triangle data as well as
the cluster data to be able to estimate the difference.

This change implements an initial basic SAH builder; we do not need to
preserve the entire tree so we just use the recursive splits to
calculate the SAH cost. The builder is not fully optimized yet, which
will happen separately.

For now all of this is in "nanite" demo; the code currently is
completely separate but it might be necessary in the future to estimate
the SAH overhead at every level of the LOD hierarchy.