demo: Extend SAH cost calculation with cluster SAH
When clusters are used for RT, we end up with an extra middle-level
structure used for ray traversal; any ray pays SAH cost for two
structures now. This change expands the computation to approximate that,
building a per-cluster BVH as well as a BVH for all cluster boxes.

There's some slop in accounting for SAH due to how sahCost functions
internally; this gives comparable results to doing the same work with
something like tiny_bvh, but should report a generally higher overhead
vs "real" hardware.