spatialorder: Improve position precision for spatialSort
Previously, we've used a three-pass radix sort to sort points or
triangles by making a 30-bit Morton code, with 10 bits per axis.

For some dense meshes, this resulted in an order that was insufficiently
precise at lower levels: positions that would fall into the same
quantization bucket would get the same Morton code and may not be
ordered properly.

Depending on the use case, this can regress various metrics, such as
compression size (if spatial sorting is used to improve compression), or
BVH quality or build performance (if spatial sorting is used as a
starting order for a BVH builder).

This change expands the precision to 16 bits per axis, which gives us 48
bits total and requires 5 radix sort passes. As a result, sorting is
~50% more expensive (but still pretty fast).

In the future we could address performance in a variety of ways, such as
switching to counting sort to avoid redundant sorts when the higher bits
are sufficiently descriptive, or using 5 9-bit passes which improves
cache coherence (at a small loss of precision, as it would require 45-bit codes).

Co-authored-by: pm4rtx <pm4rtx@gmail.com>