clusterizer: Implement initial BVH clusterizer
In order to improve performance of clusters used for raytracing, we need
to generate clusters in a SAH optimal manner; the best approach seems to
be to build a BVH and extract subtrees into clusters.

This change implements the initial version of this algorithm; as
implemented, the output clusters are not yet guaranteed to conform to
the requirements, and there will be cases where the output buffer may
overflow. This is also missing a lot of code cleanup, optimizations, and
metrics improvements.