clusterizer: Separate BVH splits from meshlet fill
Instead of appending meshlets during BVH generation, we record the
boundaries of BVH splits and then fill meshlets from the resulting
order/boundary data as a post-process.

In addition to this being cleaner, in the future that will allow us to
be a little bit more loose wrt splitting logic and falling back to more
aggressive meshlet aggregation if the generated number of meshlets is
too high as a result.