clusterizer: Replace std::sort with radix sort
Instead of using comparison sorts to sort the triangles for each axis,
use a 3-pass 10-bit radix sort; we drop the low 2 bits so that we don't
have to increase the radix bits to 11; the sorting criteria is
approximate anyway.

This also allows us to remove BVHBox::pos which improves cache locality
for subsequent access.

Together, these two optimizations make clusterization ~40% faster
without affecting the quality of the results.