clusterizer: Switch the loop order to evaluate costs per axis
Instead of aggregating all three axes at once and evaluating all three
axes at once for every split position, do this in reverse.

This results in better codegen, cleaner code, and will be more
reasonable wrt memory use in the future: right now we could remove the
3-axis allocation for costs as we only need one at a time. This change
does *not* do that, because this part of the code may need further
changes to support overlaps and other metric changes.