clusterizer: Limit BVH build depth to avoid stack overflow
While in practice the maximum depth we observe is around 30 even for
meshes with hundreds of millions of triangles, some invalid input
configurations, like NaN boxes, might result in degenerate subdivisions
that would make the algorithm N^2 and the stack consumption - O(N).

Instead, limit the tree depth arbitrarily to 50; that amounts to around
10KB of stack space in optimized builds and 20KB in debug builds so
should provide room for growth.

When this condition is hit, for now we return the range as is to the
caller, which will split it into meshlets automatically.