clusterizer: Refactor bvhSplit into more functions
Split bvhPartition and bvhFindSplit from the rest of the flow; this
makes the main splitting algorithm easier to follow and analyze in the
profiler.

As part of this, instead of accumulating and computing costs for all
three axes, we do this one axis at a time. This reduces our scratch
space consumption and should be easier to understand. There does not
seem to be a different wrt performance between these two approaches.