clusterizer: Ensure that all split clusters fit
Instead of relying on the caller to split overlong tails, we now
subdivide them during splitting; to handle arbitrary limits we try to
pack max_triangles triangles per chunk and dynamically fall back to
max_vertices / 3 if that fails. This code path should almost never be
exercised so it doesn't need to be completely optimal.

As a result, tail processing may generate clusters that are too small,
but this is a problem we already have in splitting and something that
the reassembly code will need to deal with.