clusterizer: Repack meshlets when their number exceeds bound
Because of the extra hierarchical splits, in vertex bound cases we may
run into situations where we generate more meshlets than we are allowed
to based on buildMeshletsBound which estimates a worst case *if* every
meshlet is filled to the maximum.

In that situation, we *must* repack the meshlets to avoid memory safety
issues. To do that, we disable forced splitting until we have packed
enough meshlets that the rest will fit into the allocated space.
While this naturally regresses SAH, typically the repacked meshlets are
a small fraction of the output, and repacking is in sorted order which
results in a reasonable compromise.