clusterizer: Improve cluster sizing criteria for min<max
When min_triangles < max_triangles, we used to step in units of
min_triangles, which resulted in much fewer valid splits being
evaluated.

While we need to restrict the splits to conform to min..max range, what
we actually need here is to check if the split sequence can be
representable as a sum of numbers between min and max. When the initial
sequence is divisible like that, we'd like both splits to be; otherwise
we opt for just one being representable.

This can be checked via ceil(count / max) <= floor(count / min), but
that requires two integer divisions. We use a more efficient formula
that allows to use a single idiv instruction.

This is still too expensive in the common case though: when min * 2 <=
max, every possible split that leaves one half at >= min is valid. In
this case we avoid integer divisions altogether.

This change also improves the vertex bound splits as a byproduct:
instead of forcing an even split in count/3 chunks, our splits are more
adaptive and as such more likely to produce higher quality SAH nodes.