clusterizer: Implement fill weight heuristic
When using min..max ranges, we now can optionally adjust the split
weights to increase the cost according to the unevenness of the split;
splits that divide clusters into max_triangles multiples do not change
their costs.

Increasing the fill weight effectively increases the SAH cost along with
increasing the cluster fill, making it easier to tune the subdivision
compared to using a fixed minimum triangle number.

When clusters are vertex bound, this alignment breaks SAH optimality too
much and results in suboptimal splits so we disable it.