clusterizer: Use bounding box centroids for triangles
Instead of using triangle centroids, this change switches to centroids
of bounding boxes around triangles for the purpose of sorting the
triangles.

This matches commonly used HW implementations better and results in
cleaner boundaries between triangles on regular grids as triangles that
are part of the same quad are more likely to be on the same side of a
split.