clusterizer: Adjust alignment criteria to trigger more often
When clusters are vertex bound, before this we would always treat them
as unaligned, which made splits that left very few triangles on the
right side possible.

When we are splitting between min*2 elements and max elements, it is
always possible to split into two parts, as such there's a valid aligned
split, so we can relax the condition to check for min*2 instead of max.

That results in better splits in vertex bound configurations, gaining ~1
tri/cl on average for 64v/64t for vertex heavy meshes.