Rework README integration on web site
To make the page easier to navigate, add an explicit TOC; also remove
badges that are not relevant for the web version, which makes it easier
to read on mobile.

Unfortunately we'll need to keep this divergence wrt master forever, but
absent changes in the title content it should hopefully remain
mergeable.