indexcodec: Clarify rotation during encoding
When encoding triangles via an edge reference, the rotation is implicit
in getEdgeFifo.

When encoding triangles using three vertices, rotation plays a special
role due to how reset codes are specified: reset is specified to
function regardless of the value of fea (it's conditional on feb=fec=0).
When triangles are rotated, we ensure that if any vertex matches next,
it will be the first one; as such, if feb=0 and fec=0, it must follow
that fea=0.

This is important in case this code is changed in the future to support
rotation-less encoding: for resets to not trigger accidentally on
vertices where b & c are encoded as next and a isn't (which is
impossible when rotations are used), we would need to disable `next`
adjustment for b/c if a is not next.

For now we simply note these caveats in comments. It should be easy to
only set feb/fec to 0 if fea is already 0 if we need to in the future.