vertexcodec: Add version argument to encodeVertexBufferLevel
Instead of relying strictly on the global version set via
meshopt_encodeVertexVersion, specify the version override as an extra
argument to encodeVertexBufferLevel. When set to -1, this maintains the
current behavior, but it can be set to 0/1 explicitly.

This unlocks encoding multiple streams with different versions
concurrently; it's also somewhat sensible to bundle these two because
when version is 0, the level argument does not do anything. This API
should be forward compatible in that if vertex codec v2 ever appears,
the same approach should continue working.

Note that this is specific to vertex encoding. Index encoding currently
does not have a v2, and v0 is effectively already legacy; so a separate
function does not seem warranted for now. If index codec v2 is
introduced in the future, it would make sense to make a separate
function - it would likely also need a level= argument at that point, so
the same signature would work.