clusterizer: Improve precision of bounding spheres
When computing bounding spheres in meshopt_compute*Bounds function
family, we use Ritter's algorithm that computes extremum points on 3
cardinal axes, picks the pair with the largest distance, and uses that
as an initial guess which then expands to cover all other
points/spheres.

This is efficient and often produces spheres that are close to optimal,
but in some configurations the selected points are suboptimal which
results in a larger resulting sphere.

While asymptotically efficient algorithms like Welzl exist that can
compute an optimal sphere in O(N), in practice they may have numerical
issues and do not guarantee O(N) as a worst case complexity.

EPOS attempts to solve this by running an exact solver over a small set
of extremal points, which they produce by evaluating multiple axes.
However, that still requires an exact solver, which may still have
challenges wrt numerics and performance, as even though the set of input
points is smaller, for meshlet bounds the initial set is also not very
large and thus constant factors become important.

For now, this change attempts to improve the precision by using more
than three axes to pick extremal points and then proceeds as before.
Both (+-sqrt(3)/3) and (+-sqrt(2),0) permutations were evaluated, and it
looks like the former is practically more useful, as it adds 4 axes to
test instead of 6, and produces better results.

Since this requires evaluating more axes, the extra performance impact
is mitigated by keeping track of the min/max ranges along the axes, and
by using 3-axis selection for the cone normal, which appears to be less sensitive
to selection issues.

The result improves average meshlet bounds by up to 1-2% depending on
the mesh, although sub-1% reduction is more frequent.