extern: Support sparse accessors for cgltf_accessor_read_*
Previously we have supported sparse accessors for cgltf_accessor_unpack_floats
but not for various read_ functions. This could lead to applications using
read_ functions for specific cases and the code would work for most glTF files
but fail on files with sparse data.

This change supports sparse indices by using a binary search to find the index;
the keys in sparse accessors are guaranteed to monotonically increase. This is
still slower than a single linear pass that unpack_floats does, but will work
correctly and have acceptable performance.