gltfpack: Only enable sparse mode for non-empty indices
Point clouds use empty indices, so indices.size() would be less
than the vertex count. The actual code that would execute then is a no-op,
as it would leave sparse empty, but if the mechanism of enabling sparse
mode changes in the future this may regress, so add a condition to be
explicit.