gltfpack: Error during the CMake configuration if Basis path isn't found
Without this, we would configure the build with an invalid path and the build
will subsequently fail with "missing includes" error. This is problematic
because the semantics of the path are a little strange, as it's relative to
the root CMakeLists.txt instead of the folder the CMake configuration step
started from.