Add MESHOPT_STABLE_EXPORTS option for shared libraries
When building libmeshoptimizer.so/DLL, by default we export all symbols,
including experimental ones. This is generally fine but it does mean
that the shared library produced that way is not ABI stable; subsequent
releases may break the ABI of experimental functions.

When MESHOPT_STABLE_EXPORTS option is enabled, we now disable exporting
experimental functions, which yields an ABI stable library. Note that
this, naturally, prevents using experimental functions in a build like
that; currently this option is not compatible with gltfpack since it
uses simplifySloppy which is experimental.