partition: Copy cluster indices for further processing
To accommodate future support for spatial links, we need to stop using
the source indices as is and switched to a potentially different set of
indices. To prepare for this, this change fills a freshly allocated
buffer with the copy of cluster indices, removing the duplicates in the
process. This makes adjacency builds much faster with the existing code,
and allows further simplification in cases where used[] array was used
to filter out duplicates from the same cluster (which will be done
separately).