partition: Simplify adjacency building given unique indices
Now that the cluster indices are unique, we no longer need to handle
duplicate indices anywhere in the adjacency build, which simplifies this
code and makes it faster. Additionally, countTotal() on a single-cluster
group is now O(1) as it can simply count the indices, and we stopped
using it for multi-cluster groups so it's no longer necessary.