partition: Build spatial links between neighboring clusters
Now that clusters are spatially ordered, we can compute spatial links
between neighboring clusters. To limit the effect on topological
connectivity we only build a small number of links, and we only
establish links between clusters if their bounding spheres intersect,
to avoid grouping unrelated clusters.

The cluster links are considered to be bidirectional, so we only need to
search for neighboring clusters forward in the list.