demo: Compute average cluster group radius
In addition to other metrics, we now compute the average cluster group
radius per level. We average cluster.self.radius which matches between
all clusters of an individual group by construction, so this should be
directly influenced by the quality of the partition at each level (in
addition to the clusterization which affects partitioning of course).