partition: Implement bounds based scoring
In addition to considering the boundary size we now use bounds to
compute an extra spatial factor. The highest weight is assigned to the
merges that inflate the resulting bounding sphere the least, normalized
to the target radius.

Since the groups are still only sorted by vertex count, this is not
always sufficient as we may merge into larger groups before smaller
groups based on the vertex density difference.