spatialorder: Introduce experimental point partitioner
Instead of a global Morton sort, we use hierarchical axis subdivision
that ensures that all splits are aligned to a given boundary, leaving
leaves unsorted. This is a good fit when the resulting point cloud needs
to be processed in parallel, similarly to meshlets. Compared to Morton
sort followed by a naive bucketed split, this results in consistently
smaller bounds at some extra processing cost.

Unlike spatialSortRemap, this generates an index buffer - not a remap
table. Naturally, a new set of points can be constructed without the use
of indexing.