spatialorder: Rewrite spatialSort using isolated 10-bit radix passes
Instead of computing 5 histograms in parallel, we perform 5 passes, each
of which extracts the 10-bit segment into a linear array, and then
computes the histogram and sorts by it.

By itself this may seem redundant, as it's the same amount of
computation structured differently. However, the key aspect is that
radix passes exhibit poor locality on the key lookups, and this shrinks
the size of each element 4x (from 8 bytes to 2 bytes). The linear pass
before it doesn't have issues with cache coherency and as such runs
quickly even on large arrays.

With this approach we could still compute 5 histograms in parallel, but
doing that doesn't seem meaningfully faster and the combined code is a
little simpler.