gltfpack: Implement support for lines in OBJ files
Now that fast_obj parses line information, we can use it to split the
meshes based on the primitive type. For now we assume that the input
meshes have a reasonably small number of flips between polygons and
lines and don't attempt to parse meshes out optimally; if multiple
meshes are produced, the processing code will later merge them.