clusterizer: Add references for SAH splitting
The core algorithm is due to Ingo Wald; it's a little difficult to
source a good canonical paper here but we'll just go with the kd-tree
one which is the first description of an NlogN algorithm.

The optimized modulo based formula for divisibility criteria is due to
nv_cluster_builder which currently doesn't seem to have a published (paper) reference.