demo: Remove hash/validation code
This did not end up being particularly useful in the years this code has
existed, and it overloads what's fundamentally a demo application with
extra code that is not useful for production. If future complex
algorithms are going to be developed, fuzzing drivers would probably be
a better way to tackle testing.